Adds Travis CI integration  (#119)
This PR adds a `.travis.yml` for doing continuous integration builds of the website. Once builds are running it would be sensible to enable the branch protection that requires a PR have green CI before being merged. This will help keep pull requests from breaking the build accidentally and also gives a starting point that can be extended to run a lint of the website (e.g. to find 404's, broken content).