Only install 'htmlproofer' for CI (#122)
Until there are clear instructions for how to get a sane Ruby 2.2+ install on dev machines that will cleanly install the `html-proofer` gem it should only be required for CI.

This PR tells travis to use the `Gemfile.ci` gemfile instead of the default `Gemfile`. Only the `Gemfile.ci` has `html-proofer` included.