Remove froxlor from urlcheck ignore, add ICANN.org (#159)
This commit removes `www.froxlor.org` from the ignore list for the
website linting. Previously there was a hostname mismatch error that
has since been resolved.

This commit also adds `www.ICANN.org` to the ignore list. It's always
timing out in CI and I'm not sure why, seems to work fine when I test
manually. Similarly `www.vtex.com` is suffering a DNS failure but only
in CI so I'm adding it to the ignore list.