Replace Ruby `HTMLProofer` with Python `linkchecker`. (#187)
The `HTMLProofer` based CI started failing ~a month ago. Since none of
the maintainers of the website particularly like Ruby and we dream of
dropping Jekyll it seems prudent to change to a non-Ruby website
linter too.

This commit removes `HTMLProofer`, the rakefile, and the custom Gemfile
that was used. In its place the `.travis.yml` file is updated to install
the `linkchecker` requirements. A config file and script are placed in
`util/` to do the actual link checking.

I believe this is feature-complete with what we were checking before and
it seems to be more reliable/performant overall.