Redesign HTTPS statistics to use new Firefox Telemetry  (#234)
* Redesign HTTPS statistics to use new Firefox Telemetry

Mozilla in Firefox 58 has removed API access to release telemetry for many
probes, including HTTP_PAGELOAD_IS_SSL, but has provided a dimensional dataset
via SQL [1][2].

This patch changes acquisition of the HTTPS adoption data into two parts:

 1. The historical data, which has been shown here for Let's Encrypt for some
    time. This file was cut at the date where the new data begins, 2016-11-01.
    Note that this is the same data as
    https://ct.tacticalsecret.com/secure-pageloads.tsv but trimmed and converted
    to CSV.

 2. The current data, which is generated ~daily by Mozilla. Currently this
    dataset is not published by Mozilla on a Mozilla property (being remedied),
    but for now it can be downloaded from
    https://ct.tacticalsecret.com/current-https-adoption.csv . Operations will
    need to configure the web server to download this file at a regular
    occurrence and overwrite the one being checked in as part of this commit.

These two parts are stitched together by the `httpsPlot()` method, which is
now called via a `Promise.all()` to ensure that the two dependent fetches
both complete before we start to plot the data.

The current dataset has dimensionality for the adoption figures broken out by
operating system and country. (Note: Country is by geo-ip of the submission, not
localization.)

The graphing code has provisions within it to graph any combination of country
and operating system (see the list in `httpsPlot()`). It should be
straightforward to modify this code to add in a user interface to plot custom
graphs. `gHttpsData.countryList` and `gHttpsData.osList` are populated with
the available data for each.

Note: Not all dimensions will have data for each day. If there were too few
pings for a given OS / Country combination, there will not be data in the
dataset. This is not a bug, it is a deliberate outcome of Firefox Telemetry's
privacy policy.

Since this uses Promises and Fetch, it is incompatible with Internet Explorer.
That is probably not a concern?

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1414839
[2] https://github.com/mozilla/firefox-data-docs/pull/80

* Resolve `eslint` errors and warnings as much as possible.

* Revise https traces: Only show country breakdowns (not OS), and add Japan

* Update per @alex's review comment

* Also update the cert-timeline.tsv to currently-up-to-date for stale update protection