Correct error in unique subdomains per week (#361)
* Correct error in unique subdomains per week

This calculation is out of date; it presume the certs per registered domain per
week limit is the old value of 20; the commit
7ea2fa0f1c3db2679d52be3c5f109961ffdeb25b

	commit 7ea2fa0f1c3db2679d52be3c5f109961ffdeb25b
	Author: Josh Aas <jaas@kflag.net>
	Date:   Wed Aug 1 15:40:37 2018 -0500

		Certs per domain per week rate limit has been raised to 50.

updated the certs per registered domain per week limit, but forgot to update
this dependent calculation.

* Correct error in rate limit override example

This example is also out of date; it uses the old, recently changed value for
the certs per registered doamin per week rate limit. That limit was changed in:

		commit 7ea2fa0f1c3db2679d52be3c5f109961ffdeb25b
		Author: Josh Aas <jaas@kflag.net>
		Date:   Wed Aug 1 15:40:37 2018 -0500

				Certs per domain per week rate limit has been raised to 50.

This updates the example to match.