Remove duplicate aliases. (#505)
* Remove duplicate aliases.

This can result in a translated page grabbing the redirect from the
default English page. Also add a note to TRANSLATION.md.

This can be caught using Hugo 0.55.3+'s --path-warnings flag. I'll leave
it to another PR to upgrade the version of Hugo we use in Travis and
figure out how to fail when there are warnings.

* Revert /en/ pages.