Enable 2019 merch fundraiser (#588)
This PR embeds code onto the donation page to allow individual giving
for merchandise. It's powered by Shopify.

Changes include:

- Added fundraiser.html shortcode which contains all of the
markup/embeds for the fundraiser.  - Added _fundraiser.scss for a few
additional styles, included in main.scss - Included the shortcode on
all languages donate.html page. It's in English, was told that would
be OK.

By default, the fundraiser is hidden by a style in _fundraiser.scss,
you can change that when you want to enable the fundraiser.

Following last year's fundraiser, you'll need to update the CSP to
include Shopify's url used in the embed (https://sdks.shopifycdn.com).